# 56% speedup in annotating namespaces annotating with all at once.